move float helper to it's own file
